+++ frontend/src/components/Dash/Login.vue
@@ -1,12 +1,18 @@
 <template>
 	<div class="login">
 		<h2>Login</h2>
-		<FormKit type="form" submit-label="Login" :submit-attrs="{ 'data-type': 'primary' }" @submit="login" :actions="false">
+		<FormKit type="form" submit-label="Login" :submit-attrs="{ 'button-type': 'primary' }" @submit="login">
 			<FormKit type="text" label="Your username" v-model="logindata.username" validation="required:trim" autocomplete="username" />
-			<FormKit type="password" label="Your password" v-model="logindata.password" validation="required:trim"  autocomplete="current-password" />
-			<FormKit type="text" label="Your 2FA token" v-model="logindata.token" validation="number:required|length:6,6" v-if="requires2FA" autocomplete="one-time-code" />
+			<FormKit type="password" label="Your password" v-model="logindata.password" validation="required:trim" autocomplete="current-password" />
+			<FormKit
+				type="text"
+				label="Your 2FA token"
+				v-model="logindata.token"
+				validation="number:required|length:6,6"
+				v-if="requires2FA"
+				autocomplete="one-time-code"
+			/>
 			<a @click="lostTOTP" v-if="requires2FA" class="lostTOTP">Lost your authenticator?</a>
-			<FormKit type="submit" label="Login" primary></FormKit>
 			<p>{{ response }}</p>
 		</FormKit>
 		<a @click="gotoRegister"> Register </a>